requirements.txt
flask
requests
python-dotenv
gunicorn
schedule